+++ docs/assets/js/6820b113.5b679003.js
@@ -1,0 +1,1 @@
+(window.webpackJsonp=window.webpackJsonp||[]).push([[52],{128:function(o){o.exports=JSON.parse('{"allTagsPath":"/blog/tags","slug":"download","name":"download","count":1,"permalink":"/blog/tags/download"}')}}]);