+++ docs/assets/js/6820b113.68813066.js
@@ -1,1 +1,0 @@
-(window.webpackJsonp=window.webpackJsonp||[]).push([[52],{129:function(o){o.exports=JSON.parse('{"allTagsPath":"/blog/tags","slug":"download","name":"download","count":1,"permalink":"/blog/tags/download"}')}}]);