+++ CHANGELOG.md
@@ -19,6 +19,7 @@
 
   - [改进] 框架默认序列化应该从配置中读取，而非手动编写 [#I3P1SJ](https://gitee.com/dotnetchina/Furion/issues/I3P1SJ)
   - [改进] `SqlSugar` 拓展库，支持非泛型仓储获取上下文操作对象 [#I3PK2N](https://gitee.com/dotnetchina/Furion/issues/I3PK2N)
+  - [改进] 支持分布式内存缓存可配置化 [#I3POKD](https://gitee.com/dotnetchina/Furion/issues/I3POKD)
 
 - **文档变化**